Extract polling into the apiService
